add repackaging script
Creates a htdocs directory with three sets of autodownloadable packages
for AC: one set with full image quality, one set with jpegs reduced to
80% quality and one set with jpegs reduced to 60% quality. The htdocs
directory can then server as root directory for an AC package (mirror-)
server.

Full quality skymaps and mapmodels are checked, if the filesize is small
enough for autodownload limits. If not, the jpeg image quality is reduced
to 95%.

All zips are created with maximum compression strength.